Move NewInstructorDataRowComponent under AdminHomePageComponent
