+++ src/web/app/pages-admin/admin-home-page/admin-home-page.component.ts
@@ -2,8 +2,8 @@ import { Component } from '@angular/core';
 import { finalize } from 'rxjs/operators';
 import { AccountService } from '../../../services/account.service';
 import { JoinLink } from '../../../types/api-output';
-import { InstructorData } from '../../components/new-instructor-data-row/instructor-data';
 import { ErrorMessageOutput } from '../../error-message-output';
+import { InstructorData } from './instructor-data';
 
 /**
  * Admin home page.
