+++ src/web/app/pages-admin/admin-home-page/admin-home-page.module.ts
@@ -3,8 +3,8 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { RouterModule, Routes } from '@angular/router';
 import { AjaxLoadingModule } from '../../components/ajax-loading/ajax-loading.module';
-import { NewInstructorDataRowModule } from '../../components/new-instructor-data-row/new-instructor-data-row.module';
 import { AdminHomePageComponent } from './admin-home-page.component';
+import { NewInstructorDataRowModule } from './new-instructor-data-row/new-instructor-data-row.module';
 
 const routes: Routes = [
   {
