+++ src/web/app/pages-admin/admin-home-page/new-instructor-data-row/new-instructor-data-row.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormsModule } from '@angular/forms';
 import { By } from '@angular/platform-browser';
 import { first } from 'rxjs/operators';
-import { InstructorData } from './instructor-data';
+import { InstructorData } from '../instructor-data';
 
 import { NewInstructorDataRowComponent } from './new-instructor-data-row.component';
